gpo_local_brain/requirements.txt
python-dotenv
requests
python-docx
PyPDF2 